Fail on MongoDB connection error
Prior to this commit, the server would continue to run when a database
connection could not be established. Developers would only be made aware
of this when subsequent HTTP requests failed. Since the logic for
handling requests is both conceptually and physically distant from the
server initialization code, it could be quite challenging for new
developers to diagnose the problem.

Bring the server down if initial attempt to connect to the MongoDB
fails. This should help students recognize common mistakes (e.g.
mis-typed credentials), especially when they are first getting
accustomed to working with the command-line.