Update package to reduce warning output
In Node.js version 14, the `mongodb` package causes warnings to be
printed to the console that consumers of the package such as this
project cannot controo. Version 3.5.7 of the `mongodb` package corrected
that [1]. Update the project's dependencies to use a more recent release
and reduce warning output, limiting the confusion that non-actionable
warnings can cause.

[1] https://jira.mongodb.org/browse/NODE-2536