Use Mongoose Model for "messages" collection
Prior to this commit, the application used the `Collection` API to
manipulate the "messages" collection. This interface is not included in
the project documentation, making it difficult for students to learn
about its behavior. The documentation cites it as uncommon and
recommends an alternate pattern:

> ## Connection.prototype.collection()
>
> [...]
>
> Retrieves a collection, creating it if not cached.
>
> Not typically needed by applications. Just talk to your collection
> through your model.

https://mongoosejs.com/docs/5.x/docs/api/connection.html#connection_Connection-collection

Introduce a Mongoose Model for the "messages" collection, and update the
application to use that interface.