Silence warnings issued during application startup
Prior to this commit, the Mongoose package would produce the following
warnings whenever the application made its initial database connection
attempt:

>     (node:31717) DeprecationWarning: current URL string parser is deprecated, and will be removed in a future version. To use the new parser, pass option { useNewUrlParser: true } to MongoClient.connect.
>     (Use `node --trace-deprecation ...` to show where the warning was created)
>     (node:31717) [MONGODB DRIVER] Warning: Top-level use of w, wtimeout, j, and fsync is deprecated. Use writeConcern instead.
>     (node:31717) [MONGODB DRIVER] Warning: Current Server Discovery and Monitoring engine is deprecated, and will be removed in a future version. To use the new Server Discover and Monitoring engine, pass opt
>     ion { useUnifiedTopology: true } to the MongoClient constructor.

Although the issue is technically benign, the consistent and immediate
presence of warnings trained some students to disregard their program's
output. When their modifications to the project produced novel messages,
they were less likely to perceive them.

Update the connection request to include the recommended options. Due to
an implementation detail of the `Mongoose.prototype.connect` method [1],
this change alters the value provided to the method's callback; update
the callback to preserve the original functionality.

[1] https://github.com/Automattic/mongoose/blob/4b8e0d125aa031ab8ebe12fafde76443c732190d/lib/index.js#L346-L358