Normalize server error handling
Prior to this commit, the server logic handled errors inconsistently and
sometimes neglected to close the connection which triggered the problem.
This guided students toward inadvisable usage patterns for the
Express.js framework, and it also complicated the development experience
(since programming mistakes would sometimes cause the web browser to
wait for a response indefinitely).

Update the server's routes to consistently propagate errors to the
browser using the Express.js API designed for this purpose.