Demonstrate accessibility best practices for btn
The `aria-hidden` attribute hides elements from assistive technologies
like screen readers. Applying it to elements designed for user input
makes the application inaccessible for some users. In the context of an
educational code base, this may mislead students about accessible design
patterns.

Remove the `aria-hidden` attribute so that the elements are discernible
by those using assistive technology. Replace the `i` element with a
`button` element to communicate the potential for interaction and to
enable keyboard input. Add the `aria-label` element to describe the
effect of interacting with the element. Add some styling to minimize the
effect of these changes on the application's visual appearance.